+++ CHANGELOG.md
@@ -29,3 +29,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Downloads
 - Upload Button
 - Download Button
+- Download authentication script